Commiting

+++ src/SchoolDao/SchoolManagementDaoImpl.java
@@ -103,17 +103,17 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
 
     @Override//EXCEPTION TODO
     public void findStudentById(int id) {
-        try{
-        EntityManager em = emf.createEntityManager();
-        Student student = em.find(Student.class, id);
-        System.out.println(student);
-        }catch(Exception e){
+        try {
+            EntityManager em = emf.createEntityManager();
+            Student student = em.find(Student.class, id);
+            System.out.println(student);
+        } catch (Exception e) {
             System.out.println("NO STUDENT WITH THAT ID FOUND");
         }
 
     }
-    
-      @Override//EXCEPTION TODO
+
+    @Override//EXCEPTION TODO
     public List<Student> findStudentByName(String name) {
         EntityManager em = emf.createEntityManager();
         Query query = em.createNamedQuery("showByName", Student.class);
@@ -126,7 +126,7 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
     @Override
     public List<Student> findStudentsByCourse(int id) {
         EntityManager em = emf.createEntityManager();
-        
+
         Course course = em.find(Course.class, id);
 
         List<Education> education = course.getEducations();
@@ -138,17 +138,17 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
         for (Student student : students) {
             System.out.println(student.getName());
         }
-        
+
         System.out.println("*************");
         return students;
     }
-    
+
     @Override
-    public List<Student>findStudentsByEducation(int id) {
+    public List<Student> findStudentsByEducation(int id) {
         EntityManager em = emf.createEntityManager();
         Education education = em.find(Education.class, id);
         List<Student> list = education.getStudents();
-       
+
         return list;
     }
 
@@ -172,29 +172,29 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
         return list;
 
     }
-    
-        @Override
+
+    @Override
     public List<Teacher> findTeacherByCourse(int id) {
         EntityManager em = emf.createEntityManager();
         Course course = em.find(Course.class, id);
         List<Teacher> list = course.getTeachers();
         return list;
     }
-    
-         @Override
+
+    @Override
     public List<Teacher> findTeacherByEducation(int id) {
         EntityManager em = emf.createEntityManager();
         Education education = em.find(Education.class, id);
-        
+
         List<Course> course = education.getCourses();
         List<Teacher> teachers = new ArrayList<>();
-         for (Course course1 : course) {
-             teachers.addAll(course1.getTeachers());
-         }
-         
-         for (Teacher teacher : teachers) {
-             System.out.println(teacher.getName());
-         }
+        for (Course course1 : course) {
+            teachers.addAll(course1.getTeachers());
+        }
+
+        for (Teacher teacher : teachers) {
+            System.out.println(teacher.getName());
+        }
         return teachers;
     }
 
@@ -219,35 +219,35 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
         List<Course> list = em.createNamedQuery("showAllC", Course.class).getResultList();
         return list;
     }
-    
-        @Override
+
+    @Override
     public List<Education> showAllEducations() {
         EntityManager em = emf.createEntityManager();
         List<Education> list = em.createNamedQuery("showAllE", Education.class).getResultList();
         return list;
     }
-    
-     @Override
+
+    @Override
     public void showStudentInfo() {
         List<Student> list = showAllStudents();
-         for (Student student : list) {
-             System.out.println("ID: " + student.getId() + " " + student.getName());
-         }
+        for (Student student : list) {
+            System.out.println("ID: " + student.getId() + " " + student.getName());
+        }
     }
-    
-     @Override
+
+    @Override
     public void showTeacherInfo() {
         List<Teacher> list = showAllTeachers();
-         for (Teacher teacher : list) {
-             System.out.println("ID: " + teacher.getId()+ " " + teacher.getName());
-         }
+        for (Teacher teacher : list) {
+            System.out.println("ID: " + teacher.getId() + " " + teacher.getName());
+        }
     }
 
     @Override
     public void showEducationinfo() {
         List<Education> list = showAllEducations();
         for (Education education : list) {
-            System.out.println("ID:" + education.getId()+ " " + education.getName());
+            System.out.println("ID:" + education.getId() + " " + education.getName());
         }
     }
 
@@ -255,22 +255,22 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
     public void showCourseInfo() {
         List<Course> list = showAllCourse();
         for (Course course : list) {
-            System.out.println("ID:" + course.getId()+ " " + course.getName());
+            System.out.println("ID:" + course.getId() + " " + course.getName());
         }
     }
 
     @Override //Exception ok
     public boolean updateStudentEducation(int id, int id2) {
-        try{
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        Student student = em.find(Student.class, id);
-        Education education = em.find(Education.class, id2);
-        student.setEducation(education);
-        em.persist(student);
-        em.getTransaction().commit();
-        em.close();
-        }catch(Exception e){
+        try {
+            EntityManager em = emf.createEntityManager();
+            em.getTransaction().begin();
+            Student student = em.find(Student.class, id);
+            Education education = em.find(Education.class, id2);
+            student.setEducation(education);
+            em.persist(student);
+            em.getTransaction().commit();
+            em.close();
+        } catch (Exception e) {
             System.out.println("<<Try again>>NO SUCH STUDENT OR EDUCATION FOUND<<Try again>>");
             Ui.updateMenu();
         }
@@ -279,35 +279,34 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
 
     @Override //Exception ok
     public boolean updateRemoveTeacherCourse(int id, int id2) {
-        try{
-         EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        Teacher teacher = em.find(Teacher.class, id);
-        Course course = em.find(Course.class, id2);
-        
-        
-        course.removeTeacher(teacher);
-        em.persist(course);
-        
-        em.getTransaction().commit();
-        em.close();
-        }catch(Exception e){
+        try {
+            EntityManager em = emf.createEntityManager();
+            em.getTransaction().begin();
+            Teacher teacher = em.find(Teacher.class, id);
+            Course course = em.find(Course.class, id2);
+
+            course.removeTeacher(teacher);
+            em.persist(course);
+
+            em.getTransaction().commit();
+            em.close();
+        } catch (Exception e) {
             System.out.println("<<Try again>>NO SUCH TEACHER OR COURSE FOUND<<Try again>>");
         }
         return true;
     }
-    
-     @Override //Exception ok
+
+    @Override //Exception ok
     public boolean updateStudentsName(int id, String name) {
-        try{
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        Student student = em.find(Student.class, id);
-        student.setName(name);
-        em.persist(student);
-        em.getTransaction().commit();
-        em.close();
-        }catch (Exception e){
+        try {
+            EntityManager em = emf.createEntityManager();
+            em.getTransaction().begin();
+            Student student = em.find(Student.class, id);
+            student.setName(name);
+            em.persist(student);
+            em.getTransaction().commit();
+            em.close();
+        } catch (Exception e) {
             System.out.println("<<Try again>>NO STUDENT FOUND<<Try again>>");
             Ui.updateMenu();
         }
@@ -316,30 +315,44 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
 
     @Override //Exception ok
     public boolean updateTeachersName(int id, String name) {
-        try{
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        Teacher teacher = em.find(Teacher.class, id);
-        teacher.setName(name);
-        em.persist(teacher);
-        em.getTransaction().commit();
-        em.close();
-        }catch(Exception e){
+        try {
+            EntityManager em = emf.createEntityManager();
+            em.getTransaction().begin();
+            Teacher teacher = em.find(Teacher.class, id);
+            teacher.setName(name);
+            em.persist(teacher);
+            em.getTransaction().commit();
+            em.close();
+        } catch (Exception e) {
             System.out.println("<<Try again>>NO TEACHER FOUND<<Try again>>");
             Ui.updateMenu();
         }
         return true;
 
     }
-    
-        @Override
+
+    @Override
     public boolean removeStudent(int id) {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        Student student = em.find(Student.class, id);
-        em.remove(student);
-        em.getTransaction().commit();
-        em.close();
+        try {
+            EntityManager em = emf.createEntityManager();
+            em.getTransaction().begin();
+            Student student = em.find(Student.class, id);
+            if (student != null) {
+                System.out.println(student.getId() + " " + student.getName() + " will be removed");
+                System.out.println("Are you sure you want to proceed?");
+                int val = Ui.readNumber("1.yes or 2 no");
+                if (val == 1) {
+                    System.out.println("ID " +student.getId() + " " + student.getName() + " has been remove");
+                    em.remove(student);
+                    em.getTransaction().commit();
+                    em.close();
+                }
+            }
+
+        } catch (Exception e) {
+            System.out.println("No such Student found, Try again");
+            Ui.removeMenu();
+        }
         return true;
     }
 
@@ -365,6 +378,7 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
         return true;
 
     }
+
     @Override
     public boolean removeEducation(int id) {
         EntityManager em = emf.createEntityManager();
@@ -380,5 +394,5 @@ public class SchoolManagementDaoImpl implements SchoolManagementDao {
         em.close();
         return true;
     }
-    
+
 }
